Use Path for cross platform file operations during testing
Previously worked on the mac but not the pc. Now it should work on both.

+++ taman/src/test/java/com/caldwellsoftware/taman/TamanTest.java
@@ -7,6 +7,8 @@ package com.caldwellsoftware.taman;
 
 import java.io.File;
 import java.net.URL;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import org.apache.commons.io.FileUtils;
 import static org.testng.Assert.*;
 import org.testng.annotations.Test;
@@ -29,19 +31,25 @@ public class TamanTest {
     // public void hello() {}
 
     @org.testng.annotations.BeforeClass
+    @Test(enabled=true)
     public static void setUpClass() throws Exception {
-        File outputFolder = new File(System.getProperty("user.dir"), "/output"); 
-        FileUtils.cleanDirectory(outputFolder);
+        Path outputPath = Paths.get(System.getProperty("user.dir"), "output");
+        File outputFolder = outputPath.toFile(); 
+        if (outputFolder.exists()) {
+            FileUtils.cleanDirectory(outputFolder);
+        }
         
         URL templateURL = TamanTest.class.getResource("template.xml");
-        File templateFile = new File(templateURL.getFile());
+        File templateFile = new File(templateURL.toURI());
         ta = new TaSystem(templateFile);
     }
 
     @org.testng.annotations.AfterClass
+    @Test(enabled=true)
     public static void tearDownClass() throws Exception {
-        File outputfolder = new File(System.getProperty("user.dir"), "/output"); 
-        File outFile = new File(outputfolder, "ta-output.xml");
+        Path outputPath = Paths.get(System.getProperty("user.dir"), "output", "ta-output.xml");
+        outputPath.getParent().toFile().mkdirs();
+        File outFile = outputPath.toFile();
         ta.Write(outFile);
     }
 
@@ -53,7 +61,7 @@ public class TamanTest {
     public void tearDownMethod() throws Exception {
     }
     
-    @Test
+    @Test(enabled=true)
     public void addDeclarations(){
         ta.addVarBool("myTestBool");
         ta.addVarInt("myTestInt");
@@ -62,7 +70,7 @@ public class TamanTest {
         ta.addConstInt("myTestConst", 123);
     }
     
-    @Test
+    @Test(enabled=true)
     public void addInstance(){
         ta.addInstance("INPUT", "myTestInput", "myTestBroadcastChan", "myTestBool");
     }
